Update Discord InteraKTions to 0.0.10-SNAPSHOT, add online graphs

+++ src/main/kotlin/net/perfectdreams/pantufa/interactions/commands/ChatColorExecutor.kt
@@ -1,8 +1,8 @@
 package net.perfectdreams.pantufa.interactions.commands
 
-import net.perfectdreams.discordinteraktions.common.context.SlashCommandArguments
-import net.perfectdreams.discordinteraktions.declarations.slash.SlashCommandExecutorDeclaration
-import net.perfectdreams.discordinteraktions.declarations.slash.options.CommandOptions
+import net.perfectdreams.discordinteraktions.common.context.commands.slash.SlashCommandArguments
+import net.perfectdreams.discordinteraktions.declarations.commands.slash.SlashCommandExecutorDeclaration
+import net.perfectdreams.discordinteraktions.declarations.commands.slash.options.CommandOptions
 import net.perfectdreams.pantufa.PantufaBot
 import net.perfectdreams.pantufa.utils.Constants
 import net.perfectdreams.pantufa.utils.PantufaReply
@@ -25,9 +25,9 @@ class ChatColorExecutor(pantufa: PantufaBot) : PantufaInteractionCommand(pantufa
     }
 
     override suspend fun executePantufa(context: PantufaCommandContext, args: SlashCommandArguments) {
-        val r = args[options.red]
-        val g = args[options.green]
-        val b = args[options.blue]
+        val r = args[options.red].toInt()
+        val g = args[options.green].toInt()
+        val b = args[options.blue].toInt()
 
         if (r !in 0..255 || g !in 0..255 || b !in 0..255) {
             context.reply(
