+++ src/main/kotlin/net/perfectdreams/pantufa/interactions/commands/MinecraftUserDiscordUserExecutor.kt
@@ -1,16 +1,10 @@
 package net.perfectdreams.pantufa.interactions.commands
 
-import net.perfectdreams.discordinteraktions.common.context.SlashCommandArguments
-import net.perfectdreams.discordinteraktions.declarations.slash.SlashCommandExecutorDeclaration
-import net.perfectdreams.discordinteraktions.declarations.slash.options.CommandOptions
+import net.perfectdreams.discordinteraktions.common.context.commands.slash.SlashCommandArguments
+import net.perfectdreams.discordinteraktions.declarations.commands.slash.SlashCommandExecutorDeclaration
+import net.perfectdreams.discordinteraktions.declarations.commands.slash.options.CommandOptions
 import net.perfectdreams.pantufa.PantufaBot
-import net.perfectdreams.pantufa.api.commands.SilentCommandException
-import net.perfectdreams.pantufa.dao.CashInfo
-import net.perfectdreams.pantufa.interactions.commands.MinecraftUserPlayerNameExecutor.Companion.Options.register
-import net.perfectdreams.pantufa.network.Databases
-import net.perfectdreams.pantufa.utils.Constants
 import net.perfectdreams.pantufa.utils.PantufaReply
-import org.jetbrains.exposed.sql.transactions.transaction
 
 class MinecraftUserDiscordUserExecutor(pantufa: PantufaBot) : PantufaInteractionCommand(
     pantufa
@@ -27,7 +21,7 @@ class MinecraftUserDiscordUserExecutor(pantufa: PantufaBot) : PantufaInteraction
     override suspend fun executePantufa(context: PantufaCommandContext, args: SlashCommandArguments) {
         val discordUser = args[options.user]
 
-        val minecraftUser = pantufa.getDiscordAccountFromId(discordUser.id.value) ?: run {
+        val minecraftUser = pantufa.getDiscordAccountFromId(discordUser.id.value.toLong()) ?: run {
             context.reply(
                 PantufaReply(
                     "O usuário <@${discordUser.id.value}> não tem uma conta associada!"
